frontend/requirements.txt
# Frontend dependencies
streamlit
pandas==1.3.3
pillow==8.3.2
